3.7 update

+++ vice/src/core/flash040core.c
@@ -73,7 +73,7 @@ static const flash_types_t flash_types[FLASH040_TYPE_NUM] = {
       0x80000,
       0x70000, 0x10000, 16,
       0x5555, 0x2aaa, 0x7fff, 0x7fff,
-      0x40, 
+      0x40,
       80, 2000000, 14000000}, /* may take up to 30s and 120s */
     /* AM29F040B */
     { 0x01, 0xa4, 1,
@@ -184,7 +184,7 @@ inline static int flash_program_byte(flash040_context_t *flash040_context, unsig
 inline static int flash_write_operation_status(flash040_context_t *flash040_context)
 {
     return ((flash040_context->program_byte ^ 0x80) & 0x80)   /* DQ7 = inverse of programmed data */
-           | ((maincpu_clk & 2) << 5)                         /* DQ6 = toggle bit (2 us) */
+           | ((int)(maincpu_clk & 2) << 5)                    /* DQ6 = toggle bit (2 us) */
            | (1 << 5)                                         /* DQ5 = timeout */
     ;
 }
